+++ examples/fetch_and_cat.rs
@@ -2,7 +2,10 @@ use clap::Parser;
 use futures::pin_mut;
 use futures::stream::StreamExt;
 use rust_ipfs::p2p::PeerInfo;
-use rust_ipfs::{Ipfs, IpfsPath, Multiaddr, UninitializedIpfs};
+use rust_ipfs::{Ipfs, IpfsPath, Multiaddr};
+
+use rust_ipfs::UninitializedIpfsNoop as UninitializedIpfs;
+
 use std::process::exit;
 use tokio::io::AsyncWriteExt;